[REFACTOR] Split tests for core mapping functionality and currency handling in InvoiceMapperService
