--- backend/tests/services/invoiceMapperService.test.js
@@ -1,764 +1,0 @@
-const { AzureInvoiceMapper } = require('../../src/services/invoiceMapperService');
-const { mockAzureOcrResult } = require('../mocks/azure-ocr-result');
-
-// Mock Invoice model for testing
-jest.mock('../../src/models', () => ({
-  Invoice: {
-    build: jest.fn().mockImplementation(data => data)
-  },
-  sequelize: {}
-}));
-
-describe('AzureInvoiceMapper', () => {
-  let mapper;
-  const partnerId = "contoso-partner";
-
-  beforeEach(() => {
-    mapper = new AzureInvoiceMapper();
-    jest.clearAllMocks();
-
-    // Add generatePartnerId method for partner ID tests
-    mapper.generatePartnerId = function(vendorName) {
-      if (!vendorName) return 'unknown-vendor';
-
-      let partnerId = vendorName
-        .toLowerCase()
-        .trim()
-        .replace(/\s+/g, '-')
-        .replace(/[^a-z0-9-]/g, '');
-
-      if (partnerId.length > 44) {
-        partnerId = partnerId.substring(0, 44);
-      }
-
-      return partnerId;
-    };
-  });
-
-  describe('Core Mapping Functionality', () => {
-    it('should map Azure OCR result to Invoice model format', () => {
-      const ocrResult = mockAzureOcrResult();
-      const { invoiceData } = mapper.mapToInvoiceModel(ocrResult, partnerId);
-
-      expect(invoiceData).toBeDefined();
-      expect(invoiceData.purchase_order_id).toBe(12345);
-      expect(invoiceData.invoice_date).toEqual(new Date('2023-05-15'));
-      expect(invoiceData.due_date).toEqual(new Date('2023-06-15'));
-      expect(invoiceData.total_amount).toBe(110);
-      expect(invoiceData.subtotal_amount).toBe(100);
-      expect(invoiceData.discount_amount).toBe(5);
-      expect(invoiceData.payment_terms).toBe('Null');
-      expect(invoiceData.status).toBe('Analyzed');
-      expect(invoiceData.partner_id).toBe(partnerId);
-    });
-
-    it('should handle missing fields with default values', () => {
-      const partialOcrResult = {
-        documents: [{
-          fields: {
-            InvoiceId: { content: 'INV-2023-001' },
-            InvoiceDate: { content: '2023-05-15' },
-            VendorName: { content: 'ABC Corp' },
-            InvoiceTotal: { content: '$500.00' }
-          }
-        }]
-      };
-
-      const { invoiceData } = mapper.mapToInvoiceModel(partialOcrResult, "abc-corp-partner");
-
-      expect(invoiceData.invoice_date).toEqual(new Date('2023-05-15'));
-      expect(invoiceData.total_amount).toBe(500);
-      expect(invoiceData.subtotal_amount).toBe(500);
-      expect(invoiceData.purchase_order_id).toBe(null);
-      expect(invoiceData.payment_terms).toBe(null);
-      expect(invoiceData.status).toBe('Analyzed');
-      expect(invoiceData.partner_id).toBe("abc-corp-partner");
-    });
-
-    it('should handle document without fields property', () => {
-      const ocrWithoutFields = { documents: [{}] };
-      const { invoiceData } = mapper.mapToInvoiceModel(ocrWithoutFields, partnerId);
-      
-      expect(invoiceData).toBeDefined();
-      expect(invoiceData.partner_id).toBe(partnerId);
-    });
-    
-    it('should use totalAmount as fallback when parseCurrency returns falsy for SubTotal', () => {
-      // Override parseCurrency just for this test
-      const originalParseCurrency = mapper.parseCurrency;
-      mapper.parseCurrency = jest.fn((field) => {
-        if (field === 'SubTotal') {
-          return null; // Return falsy value specifically for SubTotal
-        } else {
-          return { amount: 850, currency: { currencySymbol: '$', currencyCode: 'USD' } };
-        }
-      });
-
-      const ocrWithForcedFallback = {
-        documents: [{
-          fields: {
-            InvoiceId: { content: 'INV-FORCED-FALLBACK' },
-            InvoiceTotal: 'total', // Not used directly, our mock will handle this
-            SubTotal: 'SubTotal'   // Specific value our mock will check for
-          }
-        }]
-      };
-
-      const { invoiceData } = mapper.mapToInvoiceModel(ocrWithForcedFallback, partnerId);
-      
-      // Verify our mock was called correctly
-      expect(mapper.parseCurrency).toHaveBeenCalledWith('SubTotal');
-      expect(invoiceData.subtotal_amount).toBe(850); // Should use total_amount via direct fallback
-
-      // Restore original method
-      mapper.parseCurrency = originalParseCurrency;
-    });
-
-    it('should use totalAmountCurrency as fallback when subtotalAmount has null currency', () => {
-      // Override parseCurrency just for this test
-      const originalParseCurrency = mapper.parseCurrency;
-      mapper.parseCurrency = jest.fn((field) => {
-        if (field === 'SubTotal') {
-          // Return amount but with null currency
-          return { amount: 750, currency: null };
-        } else if (field === 'InvoiceTotal' || field === 'Total') {
-          // Return both amount and currency for total
-          return { amount: 850, currency: { currencySymbol: '€', currencyCode: 'EUR' } };
-        } else {
-          return { amount: null, currency: null };
-        }
-      });
-
-      const ocrWithCurrencyFallback = {
-        documents: [{
-          fields: {
-            InvoiceId: { content: 'INV-CURRENCY-FALLBACK' },
-            InvoiceTotal: 'total', 
-            SubTotal: 'subtotal'
-          }
-        }]
-      };
-
-      const { invoiceData } = mapper.mapToInvoiceModel(ocrWithCurrencyFallback, partnerId);
-      
-      // Verify currency fallback worked
-      expect(invoiceData.currency_symbol).toBe(null);
-      expect(invoiceData.currency_code).toBe(null);
-      
-      // Restore original method
-      mapper.parseCurrency = originalParseCurrency;
-    });
-    it('should default to null for missing currencySymbol and currencyCode', () => {
-      // Create a field with amount but no currency properties
-      const fieldWithAmountOnly = {
-        value: {
-          amount: 500
-          // No currencySymbol or currencyCode
-        }
-      };
-      
-      // Parse the currency field
-      const result = mapper.parseCurrency(fieldWithAmountOnly);
-      
-      // Verify amount was parsed correctly
-      expect(result.amount).toBe(500);
-      
-      // Verify currency properties default to null
-      expect(result.currency.currencySymbol).toBe(null);
-      expect(result.currency.currencyCode).toBe(null);
-    });
-    it('should handle empty content but non-null field in parseCurrency', () => {
-      // Create a field that has a structure but getFieldContent would return falsy value
-      const fieldWithEmptyContent = {
-        content: '',  // Empty content
-        someProperty: 'test'  // The field itself isn't null
-      };
-      
-      // Parse the currency field
-      const result = mapper.parseCurrency(fieldWithEmptyContent);
-      
-      // Verify the early return result structure is correct
-      expect(result.amount).toBeNull();
-      expect(result.currency.currencySymbol).toBeNull();
-      expect(result.currency.currencyCode).toBeNull();
-    });
-    it('should handle fileUrl in processInvoiceData method', async () => {
-      const ocrResult = mockAzureOcrResult();
-      const fileUrl = 'https://example.com/invoices/test.pdf';
-
-      const { invoiceData } = await mapper.processInvoiceData(ocrResult, partnerId, fileUrl);
-      expect(invoiceData.file_url).toBe(fileUrl);
-    });
-
-    it('should handle null fileUrl in processInvoiceData method', async () => {
-      const ocrResult = {
-        documents: [{
-          fields: {
-            InvoiceId: { content: 'INV-TEST-NULL-URL' },
-            InvoiceDate: { content: '2023-06-01' },
-            InvoiceTotal: { content: '100.00' }
-          }
-        }]
-      };
-      
-      const { invoiceData } = await mapper.processInvoiceData(ocrResult, partnerId, null);
-      expect(invoiceData.file_url).toBeNull();
-      expect(invoiceData.invoice_number).toBe('INV-TEST-NULL-URL');
-      expect(invoiceData.partner_id).toBe(partnerId);
-      
-      // Test omitting fileUrl parameter
-      const { invoiceData: invoiceData2 } = await mapper.processInvoiceData(ocrResult, partnerId);
-      expect(invoiceData2.file_url).toBeNull();
-    });
-  });
-
-  describe('Input Validation', () => {
-    it('should throw error when OCR result is invalid', () => {
-      expect(() => mapper.mapToInvoiceModel(null, partnerId))
-        .toThrow('Invalid OCR result format');
-      
-      expect(() => mapper.mapToInvoiceModel({}, partnerId))
-        .toThrow('Invalid OCR result format');
-      
-      expect(() => mapper.mapToInvoiceModel({ documents: [] }, partnerId))
-        .toThrow('Invalid OCR result format');
-    });
-    
-    it('should throw error when partnerId is missing', () => {
-      const ocrResult = mockAzureOcrResult();
-      
-      expect(() => mapper.mapToInvoiceModel(ocrResult))
-        .toThrow('Partner ID is required');
-      
-      expect(() => mapper.mapToInvoiceModel(ocrResult, null))
-        .toThrow('Partner ID is required');
-      
-      expect(() => mapper.mapToInvoiceModel(ocrResult, ''))
-        .toThrow('Partner ID is required');
-    });
-  });
-
-  describe('Date Processing', () => {
-    it('should handle invalid date format with fallback to current date', () => {
-      const invalidDate = { content: 'invalid-date' };
-      
-      const now = new Date();
-      const result = mapper.parseDate(invalidDate);
-      
-      expect(result instanceof Date).toBe(true);
-      const timeDiff = Math.abs(result.getTime() - now.getTime());
-      expect(timeDiff).toBeLessThan(1000); // Within 1 second
-    });
-    
-    it('should calculate due date based on payment terms', () => {
-      const invoiceDate = new Date('2023-05-15');
-      
-      // Various payment term formats
-      expect(mapper.calculateDueDate(invoiceDate, 'Net 30')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, 'Upon Receipt')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, '')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, 'net 45')).toEqual(new Date('2023-06-29'));
-      expect(mapper.calculateDueDate(invoiceDate, '60 days')).toEqual(new Date('2023-07-14'));
-      expect(mapper.calculateDueDate(invoiceDate, '15')).toEqual(new Date('2023-05-30'));
-      expect(mapper.calculateDueDate(invoiceDate, '  25  ')).toEqual(new Date('2023-06-09'));
-    });
-
-    it('should handle NaN, zero or negative term days with fallback to 30 days', () => {
-      const invoiceDate = new Date('2023-05-15');
-      
-      expect(mapper.calculateDueDate(invoiceDate, 'Net')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, 'n/a')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, 'Net 0')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, 'Net -10')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, '0')).toEqual(new Date('2023-06-14'));
-      expect(mapper.calculateDueDate(invoiceDate, '-5')).toEqual(new Date('2023-06-14'));
-    });
-
-    it('should default to current date when date field exists but is empty', () => {
-      const ocrWithEmptyDate = {
-        documents: [{
-          fields: {
-            InvoiceDate: { content: '' },
-            InvoiceTotal: { content: '$100.00' }
-          }
-        }]
-      };
-      
-      const now = new Date();
-      const { invoiceData } = mapper.mapToInvoiceModel(ocrWithEmptyDate, partnerId);
-      
-      const timeDiff = Math.abs(invoiceData.invoice_date.getTime() - now.getTime());
-      expect(timeDiff).toBeLessThan(1000);
-    });
-    
-  });
-  describe("Date Format Parsing", () => {
-    it('should correctly handle DD/MM/YY date format', () => {
-      const mapper = new AzureInvoiceMapper();
-      
-      // Test DD/MM/YY format (28/02/25 should be Feb 28, 2025)
-      const dateDDMMYY = mapper.parseDate({ content: '28/02/25' });
-      expect(dateDDMMYY.getFullYear()).toBe(2025);
-      expect(dateDDMMYY.getMonth() + 1).toBe(2); // Month is 0-indexed
-      expect(dateDDMMYY.getDate()).toBe(28);
-      
-      // Test another DD/MM/YY format with single digits
-      const dateSingleDigits = mapper.parseDate({ content: '5/7/23' });
-      expect(dateSingleDigits.getFullYear()).toBe(2023);
-      expect(dateSingleDigits.getMonth() + 1).toBe(7);
-      expect(dateSingleDigits.getDate()).toBe(5);
-      
-      // Test year interpretation (years < 50 are 20xx, >= 50 are 19xx)
-      const dateOlderYear = mapper.parseDate({ content: '15/06/55' });
-      expect(dateOlderYear.getFullYear()).toBe(1955);
-      expect(dateOlderYear.getMonth() + 1).toBe(6);
-      expect(dateOlderYear.getDate()).toBe(15);
-    });
-    
-    it('should correctly handle DD/MM/YYYY date format', () => {
-      const mapper = new AzureInvoiceMapper();
-      
-      // Test DD/MM/YYYY format
-      const dateDDMMYYYY = mapper.parseDate({ content: '28/02/2025' });
-      expect(dateDDMMYYYY.getFullYear()).toBe(2025);
-      expect(dateDDMMYYYY.getMonth() + 1).toBe(2);
-      expect(dateDDMMYYYY.getDate()).toBe(28);
-      
-      // Test with single digits
-      const dateSingleDigits = mapper.parseDate({ content: '5/7/2023' });
-      expect(dateSingleDigits.getFullYear()).toBe(2023);
-      expect(dateSingleDigits.getMonth() + 1).toBe(7);
-      expect(dateSingleDigits.getDate()).toBe(5);
-    });
-    
-    it('should correctly handle invalid date formats by using current date', () => {
-      const mapper = new AzureInvoiceMapper();
-      
-      // Mock date.now to have consistent test results
-      const now = new Date('2025-01-01T12:00:00Z');
-      jest.spyOn(global, 'Date').mockImplementation(() => now);
-      
-      // Test invalid format that should fall back to current date
-      const invalidDate = mapper.parseDate({ content: 'not-a-date' });
-      expect(invalidDate).toEqual(now);
-      
-      // Clean up
-      jest.restoreAllMocks();
-    });
-  });
-  describe('Field Parsing', () => {
-    it('should safely handle field content extraction', () => {
-      expect(mapper.getFieldContent(null)).toBeNull();
-      expect(mapper.getFieldContent(undefined)).toBeNull();
-      expect(mapper.getFieldContent({})).toBeNull();
-      expect(mapper.getFieldContent({ content: 'test' })).toBe('test');
-      
-      const fieldWithDirectValue = { value: 'direct value', content: null };
-      expect(mapper.getFieldContent(fieldWithDirectValue)).toBe('direct value');
-      
-      const fieldWithTextInValue = { value: { text: 'text in value object' }, content: null };
-      expect(mapper.getFieldContent(fieldWithTextInValue)).toBe('text in value object');
-    });
-    
-    it('should return null for fields with unexpected data structure', () => {
-      const unexpectedStructure = { 
-        someProperty: 'value',
-        nestedObject: { anotherProperty: 'another value' }
-      };
-      
-      expect(mapper.getFieldContent(unexpectedStructure)).toBeNull();
-    });
-    
-    it('should safely handle numeric parsing', () => {
-      expect(mapper.parseNumeric(null)).toBeNull();
-      expect(mapper.parseNumeric({ content: 'abc' })).toBeNull();
-      expect(mapper.parseNumeric({ content: '123' })).toBe(123);
-      expect(mapper.parseNumeric({ content: '123.45' })).toBe(123.45);
-      
-      // Direct numeric value tests
-      expect(mapper.parseNumeric({ value: 42 })).toBe(42);
-      expect(mapper.parseNumeric({})).toBeNull();
-      expect(mapper.parseNumeric({ value: "a" })).toBeNull();
-    });
-    
-    it('should explicitly handle all branches of direct numeric value check', () => {
-      expect(mapper.parseNumeric(null)).toBeNull();
-      expect(mapper.parseNumeric(undefined)).toBeNull();
-      expect(mapper.parseNumeric({})).toBeNull();
-      
-      const nonNumberValues = [
-        { value: 'tes' },
-        { value: true },
-        { value: [] },
-        { value: {} },
-        { value: null }
-      ];
-      
-      nonNumberValues.forEach(field => {
-        expect(mapper.parseNumeric(field)).toBeNull();
-      });
-      
-      expect(mapper.parseNumeric({ value: 42 })).toBe(42);
-      expect(mapper.parseNumeric({ value: 0 })).toBeNull();
-      expect(mapper.parseNumeric({ value: -1.5 })).toBe(-1.5);
-    });
-    
-    it('should parse currency values correctly', () => {
-      // Test cases for string content
-      expect(mapper.parseCurrency({ content: '$123.45' })).toEqual({ amount: 123.45, currency: { currencySymbol: '$', currencyCode: null } });
-      expect(mapper.parseCurrency({ content: '£123.45' })).toEqual({ amount: 123.45, currency: { currencySymbol: '£', currencyCode: null } });
-      expect(mapper.parseCurrency({ content: '123.45 EUR' })).toEqual({ amount: 123.45, currency: { currencySymbol: null, currencyCode: null } });
-      expect(mapper.parseCurrency({ content: 'abc' })).toEqual({ amount: null, currency: { currencySymbol: null, currencyCode: null } });
-    
-      // Test case for structured currency field
-      const structuredCurrencyField = {
-        value: {
-          amount: 129.99,
-          currencyCode: 'USD',
-          currencySymbol: '$'
-        }
-      };
-      expect(mapper.parseCurrency(structuredCurrencyField)).toEqual({ amount: 129.99, currency: { currencySymbol: '$', currencyCode: 'USD' } });
-    
-      // Test case for direct number value
-      expect(mapper.parseCurrency({ value: 42 })).toEqual({ amount: 42, currency: { currencySymbol: null, currencyCode: null } });
-    
-      // Test case for null field
-      expect(mapper.parseCurrency(null)).toEqual({ amount: null, currency: { currencySymbol: null, currencyCode: null } });
-    });
-    it('should correctly parse Rupiah currency format', () => {
-      const mapper = new AzureInvoiceMapper();
-      
-      // Test structured currency field with Rupiah content
-      const rupiahField = {
-        value: {
-          amount: 100000, // This value should be overridden by the parsed content
-          currencySymbol: '$', // This should be replaced with Rp
-          currencyCode: 'USD' // This should be replaced with IDR
-        },
-        content: 'Rp67.998'
-      };
-      
-      const result = mapper.parseCurrency(rupiahField);
-      
-      // Should parse Indonesian number format (67.998 → 67998)
-      expect(result.amount).toBe(67998);
-      expect(result.currency.currencySymbol).toBe('Rp');
-      expect(result.currency.currencyCode).toBe('IDR');
-      
-      // Test with decimal comma
-      const rupiahWithDecimal = {
-        value: {
-          amount: 50000
-        },
-        content: 'Rp45.750,50'
-      };
-      
-      const resultWithDecimal = mapper.parseCurrency(rupiahWithDecimal);
-      // 45.750,50 should be converted to 45750.50
-      expect(resultWithDecimal.amount).toBe(45750.50);
-      expect(resultWithDecimal.currency.currencySymbol).toBe('Rp');
-      expect(resultWithDecimal.currency.currencyCode).toBe('IDR');
-      
-      // Test with no thousands separator
-      const simplifiedRupiah = {
-        value: {
-          amount: 1000
-        },
-        content: 'Rp5000'
-      };
-      
-      const simpleResult = mapper.parseCurrency(simplifiedRupiah);
-      expect(simpleResult.amount).toBe(5000);
-      expect(simpleResult.currency.currencySymbol).toBe('Rp');
-      expect(simpleResult.currency.currencyCode).toBe('IDR');
-    });
-    it('should handle edge cases in parsePurchaseOrderId', () => {
-      expect(mapper.parsePurchaseOrderId({ content: '12345' })).toBe(12345);
-      expect(mapper.parsePurchaseOrderId({ content: 'ABC-DEF' })).toBe(0);
-      expect(mapper.parsePurchaseOrderId({ content: '-' })).toBe(0);
-      expect(mapper.parsePurchaseOrderId({ content: '12345e678' })).toBe(12345678);
-      expect(mapper.parsePurchaseOrderId({ content: 'Infinity' })).toBe(0);
-      expect(mapper.parsePurchaseOrderId({ content: 'NaN' })).toBe(0);
-    });
-    
-    it('should handle tax field variations', () => {
-      const ocrWithTaxVariations = {
-        documents: [{
-          fields: {
-            TotalTax: { content: '15.00' },
-            Tax: { content: '10.00' }
-          }
-        }]
-      };
-      
-      const { invoiceData } = mapper.mapToInvoiceModel(ocrWithTaxVariations, partnerId);
-      expect(invoiceData.tax_amount).toBe(15); // Should use TotalTax first
-    });
-  });
-
-  describe('Line Item Processing', () => {
-    
-    it('should handle various line item formats', () => {
-      // Standard format
-      const standardItems = {
-        values: [
-          {
-            properties: {
-              Description: { content: 'Standard Item' },
-              Quantity: { content: '2' },
-              UnitPrice: { content: '10.00' },
-              Amount: { content: '20.00' }
-            }
-          }
-        ]
-      };
-      
-      const result1 = mapper.extractLineItems(standardItems);
-      expect(result1[0].description).toBe('Standard Item');
-      expect(result1[0].quantity).toBe(2);
-      expect(result1[0].unitPrice).toBe(10);
-      expect(result1[0].amount).toBe(20);
-      
-      // Alternative field names
-      const alternativeItems = {
-        values: [
-          {
-            properties: {
-              ProductCode: { content: 'Alternative field name' },
-              Quantity: { content: '2' },
-              UnitPrice: { content: '45.00' },
-              Amount: { content: '90.00' }
-            }
-          }
-        ]
-      };
-      
-      const result2 = mapper.extractLineItems(alternativeItems);
-      expect(result2[0].description).toBe('Alternative field name');
-      expect(result2[0].quantity).toBe(2);
-      expect(result2[0].unitPrice).toBe(45);
-      expect(result2[0].amount).toBe(90);
-      
-      // Missing fields
-      const partialItems = {
-        values: [
-          {
-            properties: {
-              Description: null,
-              Quantity: { content: '3' },
-              Amount: { content: '45.00' }
-            }
-          }
-        ]
-      };
-      
-      const result3 = mapper.extractLineItems(partialItems);
-      expect(result3[0].description).toBeNull();
-      expect(result3[0].quantity).toBe(3);
-      expect(result3[0].unitPrice).toBeNull();
-      expect(result3[0].amount).toBe(45);
-      
-      // Missing properties
-      const itemsWithMissingproperties = {
-        values: [{ someOtherProperty: 'test' }]
-      };
-      
-      const result4 = mapper.extractLineItems(itemsWithMissingproperties);
-      expect(result4).toHaveLength(1);
-      expect(result4[0].description).toBeNull();
-      expect(result4[0].quantity).toBeNull();
-    });
-    
-    it('should handle non-structured line items', () => {
-      const itemsWithContent = {
-        content: 'Single line item description without structured data'
-      };
-      
-      const result = mapper.extractLineItems(itemsWithContent);
-      expect(result).toHaveLength(1);
-      expect(result[0].description).toBe('Single line item description without structured data');
-      expect(result[0].quantity).toBeNull();
-      expect(result[0].unitPrice).toBeNull();
-      expect(result[0].amount).toBeNull();
-    });
-    
-    it('should return empty array when items field has no usable content', () => {
-      const emptyContentItemsField = {
-        content: '',
-        someOtherProperty: 'value'
-      };
-      
-      expect(mapper.extractLineItems(emptyContentItemsField)).toEqual([]);
-      
-      const nullContentItemsField = {
-        someProperty: 'test',
-        content: null
-      };
-      
-      
-      expect(mapper.extractLineItems(nullContentItemsField)).toEqual([]);
-    });
-  });
-
-  describe('Partner ID Generation', () => {
-    it('should generate valid partner IDs from vendor names', () => {
-      expect(mapper.generatePartnerId('Acme Corp Ltd.')).toBe('acme-corp-ltd');
-      expect(mapper.generatePartnerId('   Spaces   ')).toBe('spaces');
-      expect(mapper.generatePartnerId('')).toBe('unknown-vendor');
-      expect(mapper.generatePartnerId(null)).toBe('unknown-vendor');
-      
-      // Test truncation for long names
-      const longName = 'Really Long Company Name That Should Be Truncated To Fit Database Fields';
-      expect(mapper.generatePartnerId(longName).length).toBeLessThanOrEqual(44);
-    });
-  });
-
-
-  describe('extractInvoiceNumber', () => {
-    test('should extract from InvoiceId field when available', () => {
-      // Arrange
-      const fields = {
-        InvoiceId: { value: 'INV-12345' },
-        InvoiceNumber: { value: 'Should not use this' },
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-
-    test('should extract from InvoiceNumber when InvoiceId is not available', () => {
-      // Arrange
-      const fields = {
-        InvoiceNumber: { value: 'INV-12345' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-
-    test('should extract from "Invoice number" when other fields are not available', () => {
-      // Arrange
-      const fields = {
-        "Invoice number": { value: 'INV-12345' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-
-    test('should extract from "Invoice #" when other fields are not available', () => {
-      // Arrange
-      const fields = {
-        "Invoice #": { value: 'INV-12345' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-
-    test('should extract from "Invoice No" when other fields are not available', () => {
-      // Arrange
-      const fields = {
-        "Invoice No": { value: 'INV-12345' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-
-    test('should extract from "Invoice No." when other fields are not available', () => {
-      // Arrange
-      const fields = {
-        "Invoice No.": { value: 'INV-12345' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-
-    test('should return empty string when no invoice number field is available', () => {
-      // Arrange
-      const fields = {
-        SomeOtherField: { value: 'Not an invoice number' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('');
-    });
-
-    test('should return empty string when fields is null or undefined', () => {
-      // Arrange - Null case
-      const nullFields = null;
-
-      // Act
-      const nullResult = mapper.extractInvoiceNumber(nullFields);
-
-      // Assert
-      expect(nullResult).toBe('');
-
-      // Arrange - Undefined case
-      const undefinedFields = undefined;
-
-      // Act
-      const undefinedResult = mapper.extractInvoiceNumber(undefinedFields);
-
-      // Assert
-      expect(undefinedResult).toBe('');
-    });
-
-    test('should handle different value formats from OCR', () => {
-      // Arrange - Field with content property
-      const fieldsWithContent = {
-        InvoiceId: { content: 'INV-12345' }
-      };
-
-      // Arrange - Field with nested text value
-      const fieldsWithNestedValue = {
-        InvoiceId: { value: { text: 'INV-67890' } }
-      };
-
-      // Act
-      const contentResult = mapper.extractInvoiceNumber(fieldsWithContent);
-      const nestedResult = mapper.extractInvoiceNumber(fieldsWithNestedValue);
-
-      // Assert
-      expect(contentResult).toBe('INV-12345');
-      expect(nestedResult).toBe('INV-67890');
-    });
-
-    test('should trim whitespace from extracted value', () => {
-      // Arrange
-      const fields = {
-        InvoiceId: { value: '  INV-12345  ' }
-      };
-
-      // Act
-      const result = mapper.extractInvoiceNumber(fields);
-
-      // Assert
-      expect(result).toBe('INV-12345');
-    });
-  });
-});
